Rename MessageListCommon -> MessageListCentrifuge
